feat: slider logic now only works on image track
